Adding base coverage data
Coverage data is now available in the variant view via variant['base_coverage']

Coverage is a list of dicts; one per base in the variant. So a 2 base insertion has 1 base, of coverage, and a 2 base deletion has 3 bases.
Dicts are of the form:
    {
	'has_coverage': True,
        'xpos': 1e9+1,
        'mean': 0.0,
        'median': 0.0,
        '1': 0.0,
        '5': 0.0,
        '10': 0.0,
        '15': 0.0,
        '20': 0.0,
        '25': 0.0,
        '30': 0.0,
        '50': 0.0,
        '100': 0.0,
    }

Or, if the base was not targeted:

    {
        'xpos': 1e9+1,
        'has_coverage': False,
    }

There is also a new config - app.config['BASE_COVERAGE_FILES']
This is a list of coverage files (eg. one per chromosome)

When you run load_db, these files are loaded into a base_coverage collection which is assumed to have one entry per base in the target region.
(though I'm not sure what 'target region' actually means here - Monkol and Konrad created these files)